commit code by abhishek
commit code by abhishek

+++ wp-content/themes/porto/inc/plugins/importer/porto-wordpress-importer.php
@@ -28,7 +28,7 @@ if ( ! class_exists( 'WP_Importer' ) ) {
 }
 
 // include WXR file parsers
-require dirname( __FILE__ ) . '/porto-parsers.php';
+require PORTO_PLUGINS . '/importer/porto-parsers.php';
 
 /**
  * WordPress Importer class for managing the import process of a WXR file
@@ -152,7 +152,7 @@ if ( class_exists( 'WP_Importer' ) ) {
 			switch ( $process ) {
 				case 'import_start':
 					// register product attributes
-					$demo        = ( isset( $_POST['demo'] ) && $_POST['demo'] ) ? $_POST['demo'] : 'landing';
+					$demo        = ( isset( $_POST['demo'] ) && $_POST['demo'] ) ? sanitize_text_field( $_POST['demo'] ) : 'landing';
 					$extra_demos = Porto_Theme_Setup_Wizard::get_instance()->porto_extra_demos();
 					if ( ! in_array( $demo, $extra_demos ) && function_exists( 'wc_get_attribute_taxonomies' ) ) {
 						$attribute_names  = array( 'Color', 'Size', 'Brand' );
@@ -200,7 +200,7 @@ if ( class_exists( 'WP_Importer' ) ) {
 							$charset_collate .= " COLLATE {$wpdb->collate}";
 						}
 
-						if ( ! $wpdb->get_var( "SHOW TABLES LIKE '$table_name'" ) == $table_name ) {
+						if ( ! $wpdb->get_var( $wpdb->prepare( 'SHOW TABLES LIKE %s', $table_name ) ) == $table_name ) {
 							$sql .= "CREATE TABLE {$table_name} (
 									meta_id bigint(20) NOT NULL AUTO_INCREMENT,
 									{$taxonomy}_id bigint(20) NOT NULL default 0,
@@ -341,6 +341,7 @@ if ( class_exists( 'WP_Importer' ) ) {
 					$posts[] = $post;
 				}
 			}
+
 			$posts = array_merge( $posts, $nav_items );
 
 			$this->version = $import_data['version'];
@@ -625,6 +626,21 @@ if ( class_exists( 'WP_Importer' ) ) {
 						$this->processed_terms[ intval( $cat['term_id'] ) ] = (int) $term_id;
 					}
 					update_option( 'porto_import_processed_terms', $this->processed_terms );
+
+					$tax_meta_array = get_metadata( 'category', $term_id );
+					if ( $tax_meta_array && is_array( $tax_meta_array ) ) {
+						foreach ( $tax_meta_array as $old_meta_key => $old_meta_value ) {
+							delete_metadata( 'category', $term_id, $old_meta_key );
+						}
+					}
+					if ( ! empty( $cat['tax_meta'] ) ) {
+						foreach ( $cat['tax_meta'] as $meta ) {
+							$key   = $meta['key'];
+							$value = $meta['value'];
+							update_metadata( 'category', $term_id, $key, $value );
+						}
+					}
+
 					continue;
 				}
 
@@ -743,6 +759,21 @@ if ( class_exists( 'WP_Importer' ) ) {
 							$this->processed_terms[ intval( $term['term_id'] ) ] = (int) $term_id;
 						}
 						update_option( 'porto_import_processed_terms', $this->processed_terms );
+
+						$tax_meta_array = get_metadata( $term['term_taxonomy'], $term_id );
+						if ( $tax_meta_array && is_array( $tax_meta_array ) ) {
+							foreach ( $tax_meta_array as $old_meta_key => $old_meta_value ) {
+								delete_metadata( $term['term_taxonomy'], $term_id, $old_meta_key );
+							}
+						}
+						if ( ! empty( $term['tax_meta'] ) ) {
+							foreach ( $term['tax_meta'] as $meta ) {
+								$key   = $meta['key'];
+								$value = $meta['value'];
+								update_metadata( $term['term_taxonomy'], $term_id, $key, $value );
+							}
+						}
+
 						continue;
 					}
 
@@ -988,13 +1019,18 @@ if ( class_exists( 'WP_Importer' ) ) {
 						delete_post_meta( $post_id, $key );
 					}
 
-					$meta_fields = porto_ct_default_view_meta_fields();
-					foreach ( $meta_fields as $key => $value ) {
-						delete_post_meta( $post_id, $key );
+					if ( function_exists( 'porto_ct_default_view_meta_fields' ) ) {
+						$meta_fields = porto_ct_default_view_meta_fields();
+						foreach ( $meta_fields as $key => $value ) {
+							delete_post_meta( $post_id, $key );
+						}
 					}
-					$meta_fields = porto_ct_default_skin_meta_fields();
-					foreach ( $meta_fields as $key => $value ) {
-						delete_post_meta( $post_id, $key );
+
+					if ( function_exists( 'porto_ct_default_skin_meta_fields' ) ) {
+						$meta_fields = porto_ct_default_skin_meta_fields();
+						foreach ( $meta_fields as $key => $value ) {
+							delete_post_meta( $post_id, $key );
+						}
 					}
 				}
 			} else {
@@ -1183,6 +1219,12 @@ if ( class_exists( 'WP_Importer' ) ) {
 				$post['postmeta'] = apply_filters( 'wp_import_post_meta', $post['postmeta'], $post_id, $post );
 
 				// add/update post meta
+			if ( 'page' == $post['post_type'] ) {
+				if ( $post_exists ) {
+					delete_post_meta( $post_id, 'porto_imported_date' );
+				}
+				add_post_meta( $post_id, 'porto_imported_date', current_time( 'mysql' ) );
+			}
 			if ( ! empty( $post['postmeta'] ) ) {
 				foreach ( $post['postmeta'] as $meta ) {
 					$key   = apply_filters( 'import_post_meta_key', $meta['key'], $post_id, $post );
@@ -1293,7 +1335,7 @@ if ( class_exists( 'WP_Importer' ) ) {
 				$_menu_item_classes = implode( ' ', $_menu_item_classes );
 			}
 
-			if ( isset( $_menu_item_url ) && ( 'http://sw-themes.com/porto_dummy/portfolio' == $_menu_item_url || 'http://sw-themes.com/porto_dummy/portfolio/' == $_menu_item_url || 'http://sw-themes.com/porto_dummy/member' == $_menu_item_url || 'http://sw-themes.com/porto_dummy/member/' == $_menu_item_url ) ) {
+			if ( isset( $_menu_item_url ) && strpos( $_menu_item_url, 'http://sw-themes.com/porto_dummy' ) === 0 ) {
 				$_menu_item_url = str_replace( 'http://sw-themes.com/porto_dummy', get_site_url(), $_menu_item_url );
 			}
